% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_simplex.R
\name{is_simplex}
\alias{is_simplex}
\title{Is Vector a Simplex?}
\usage{
is_simplex(x, K, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Vector to check.}

\item{K}{Dimension of the simplex.}

\item{tol}{Numeric tolerance for checking non-negativity and sum to 1.}
}
\description{
Checks if a numeric vector x is a valid simplex of dimension K, i.e.,
all elements are non-negative, sum to 1, and length of x is K.
}
