% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm.R
\name{scm}
\alias{scm}
\title{Robust Synthetic Control}
\usage{
scm(
  data,
  fixed_v = NULL,
  scaling = !identical(fixed_v, "uniform"),
  trials = 20,
  nloptr_control = list(),
  qp_control = list(),
  alpha = 0.1,
  synthlike_output = FALSE
)
}
\arguments{
\item{data}{Either an output from \code{\link[Synth:dataprep]{Synth::dataprep()}}, or a list with
matrix \code{Y0} and vector \code{Y1} of post-treatment outcomes of control units and
treated unit, respectively; matrix \code{X0} and vector \code{X1} of pre-treatment
predictors of control units and treated unit, respectively;
and matrix \code{Z0} and vector \code{Z1} of pre-treatment outcomes used in
optimizing donor weights w of control units and treated unit, respectively.}

\item{fixed_v}{Either \code{NULL} (default), \code{"uniform"}, or custom values.
If \code{NULL}, initializes the predictor weights uniformly. If \code{"uniform"}, uses
fixed uniform weights. If custom values, should be a numeric simplex vector
of length equal to the number of predictors which are used as fixed values for
predictor weights v.}

\item{scaling}{Logical, whether to scale each predictor to with their
pretreatment standard deviation. Default is \code{TRUE} unless \code{fixed_v} is set to
\code{"uniform"}.}

\item{trials}{Positive integer, number of random initializations to use
when \code{fixed_v} is \code{NULL}. First optimization run uses uniform weights,
while for the rest (if \code{trials} > 1) the initial values are sampled from
symmetric Dirichlet(0.5) distribution. The default is 20.}

\item{nloptr_control}{A list of control parameters to be passed to
\code{\link[nloptr:nloptr]{nloptr::nloptr()}} as argument \code{control}. See documentation of
\code{\link[nloptr:nloptr]{nloptr::nloptr()}} for details.}

\item{qp_control}{A list of control parameters to be passed to
\code{\link[limSolve:lsei]{limSolve::lsei()}}, i.e, arguments \code{type}, \code{tol}, \code{tolrank}, and \code{verbose}.}

\item{alpha}{Positive numeric scalar, concentration parameter for symmetric
Dirichlet distribution used for sampling initial predictor weights Default
is 0.1, which puts more mass on sparse predictor weights.}

\item{synthlike_output}{Logical. If \code{TRUE}, uses same output format as in
\link[Synth:synth]{Synth::synth}, i.e. the output is a list with elements such as \code{solution.v}
and \code{loss.v} instead of \code{v} and \code{loss_v}. Default is \code{FALSE}. Setting this
\code{TRUE} allows the use of \code{robustSynth} as a drop-in replacement for \code{Synth}.}
}
\description{
Estimates synthetic control using multiple initial values for predictor
weights to improve the probability of convergence to the global optimum.
This can be parallelized using \code{\link[future:plan]{future::plan()}}, with automatic progress
reporting by \code{\link[progressr:with_progress]{progressr::with_progress()}} (see examples below).
}
\details{
In contrast to the original Synth package, this implementation uses a
computationally more efficient quadratic programming of \code{\link[limSolve:lsei]{limSolve::lsei()}}
for donor weight optimization, and optimization algorithms of \code{nloptr} for
predictor weight optimization. By default, both solver types of
\code{\link[limSolve:lsei]{limSolve::lsei()}} are used and the one with better solution is picked. For
\code{\link[nloptr:nloptr]{nloptr::nloptr()}}, the default optimization algorithm is the
derivative-free \code{NLOPT_LN_SBPLX}.

Finally, whether or not to scale predictors using their pre-treatment
standard deviations can be controlled using argument \code{scaling}. This is
especially useful if you want to use all pre-treatment outcomes as
predictors with uniform weights (i.e., \code{fixed_v = "uniform"} and
\code{scaling = FALSE}).
}
\examples{
if (requireNamespace("Synth", quietly = TRUE)) {
  # from ?Synth::synth
  data(synth.data, package = "Synth")

  d <- Synth::dataprep(
    synth.data,
    predictors = c("X1", "X2", "X3"),
    predictors.op = "mean",
    dependent = "Y",
    unit.variable = "unit.num",
    time.variable = "year",
    special.predictors = list(
      list("Y", 1991, "mean"), list("Y", 1985, "mean"), list("Y", 1980, "mean")
    ),
    treatment.identifier = 7,
    controls.identifier = c(29, 2, 13, 17, 32, 38),
    time.predictors.prior = c(1984:1989),
    time.optimize.ssr = c(1984:1990),
    unit.names.variable = "name",
    time.plot = 1984:1996
   )
   synth_out <- Synth::synth(d)
   # for parallelization, use, e.g., future::plan(multisession, workers = 4)
   if (interactive()) {
     # show progress bar
     progressr::handlers(global = TRUE)
   }
   robust_out <- scm(d, trials = 20)
   synth_out$loss.v
   robust_out$loss_v
   data.frame(
     time = d$tag$time.plot,
     synth = c(d$Y1plot - d$Y0plot \%*\% synth_out$solution.w),
     robust = robust_out$effect
   )
}
}
