% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustSynth-package.R
\docType{package}
\name{robustSynth-package}
\alias{robustSynth}
\alias{robustSynth-package}
\title{The 'robustSynth' package}
\description{
Fast and robust synthetic control method analogous to Synth
package, but with improved numerical stability and computational efficiency
to ensure the returned solution is actually an optimum.
}
\author{
\strong{Maintainer}: Jouni Helske \email{jouni.helske@iki.fi}

}
